move examples to webknossos/examples and webknossos/examples/WIP
