+++ webknossos/webknossos/examples/WIP/merge_trees_at_closest_nodes.py
