+++ webknossos/webknossos/examples/WIP/offline_merger_mode.py
